+++ examples/account/account_update_transaction.py
@@ -6,35 +6,18 @@ python examples/account/account_update_transaction.py
 """
 
 import datetime
-import os
 import sys
 
-from dotenv import load_dotenv
-
-from hiero_sdk_python import AccountId, Client, Duration, Hbar, Network, PrivateKey
+from hiero_sdk_python import Client, Duration, Hbar, PrivateKey
 from hiero_sdk_python.account.account_create_transaction import AccountCreateTransaction
 from hiero_sdk_python.account.account_update_transaction import AccountUpdateTransaction
 from hiero_sdk_python.query.account_info_query import AccountInfoQuery
 from hiero_sdk_python.response_code import ResponseCode
 from hiero_sdk_python.timestamp import Timestamp
 
-load_dotenv()
 
-network_name = os.getenv("NETWORK", "testnet").lower()
-
-
-def setup_client():
-    """Initialize and set up the client with operator account"""
-    network = Network(network_name)
-    print(f"Connecting to Hedera {network_name} network!")
-    client = Client(network)
-
-    operator_id = AccountId.from_string(os.getenv("OPERATOR_ID", ""))
-    operator_key = PrivateKey.from_string(os.getenv("OPERATOR_KEY", ""))
-    client.set_operator(operator_id, operator_key)
-    print(f"Client set up with operator id {client.operator_account_id}")
-
-    return client
+def setup_client() -> Client:
+    return Client.from_env()
 
 
 def create_account(client):